requirements.txt
coreapi==2.1.1
Django==1.10.5
django-filter==1.0.1
django-macaddress==1.4.0
django-rest-swagger==2.1.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
itypes==1.1.0
Markdown==2.6.7
mock==2.0.0
netaddr==0.7.19
openapi-codec==1.2.1
pbr==1.10.0
PyJWT==1.4.2
requests==2.12.5
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
